+++ client-api/Readme.md
@@ -1,0 +1,5 @@
+# Run tesst 
+
+go test client-api/*.go -run Test_Transfer_Network  
+
+go test client-api/*.go  -run Test_RateLimit
